Moved LibraryPrep class from main.py to seperate file in data_processing
